appengine/standard_python37/hello_world/requirements.txt
responder
gunicorn
uvicorn
sklearn
google-cloud-logging